test(appsec): own gevent cleanup regressions
